use 1 app.ready hook
