+++ conf/settings.py
@@ -46,9 +46,10 @@ INSTALLED_APPS = [
     "applications.apps.ApplicationsConfig",
     "audit_trail",
     "background_task",
-    "cases.app.CasesConfig",
+    "cases",
     "cases.generated_documents",
     "compliance",
+    "conf.app.Config",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
@@ -61,7 +62,7 @@ INSTALLED_APPS = [
     "goodstype",
     "gov_users",
     "letter_templates",
-    "licences.app.LicencesConfig",
+    "licences",
     "organisations",
     "parties",
     "picklists",
