+++ licences/tasks.py
@@ -50,9 +50,9 @@ def send_licence_to_hmrc_integration(licence_id, licence_reference, scheduled_as
     """
 
     with transaction.atomic():
-        # transaction.atomic + select_for_update + nowait=True will throw an error if row has already been locked
-        licence = Licence.objects.select_for_update(nowait=True).get(id=licence_id)
         try:
+            # transaction.atomic + select_for_update + nowait=True will throw an error if row has already been locked
+            licence = Licence.objects.select_for_update(nowait=True).get(id=licence_id)
             hmrc_integration_operations.send_licence(licence)
         except hmrc_integration_operations.HMRCIntegrationException as exc:
             _handle_exception(str(exc), licence_id, licence_reference, scheduled_as_background_task)
